+++ src/coordinate/coor4d.rs
@@ -1,84 +1,24 @@
-use crate::math::angular;
-use std::ops::{Add, Div, Mul, Sub};
+use crate::coordinate::*;
 
 /// Generic 4D coordinate tuple, with no fixed interpretation of the elements
 #[derive(Debug, Default, PartialEq, Copy, Clone)]
 pub struct Coor4D(pub [f64; 4]);
 
-// ----- O P E R A T O R   T R A I T S -------------------------------------------------
-
-use std::ops::{Index, IndexMut};
-
-impl Index<usize> for Coor4D {
-    type Output = f64;
-    fn index(&self, i: usize) -> &Self::Output {
-        &self.0[i]
+impl CoordinateTuple for Coor4D {
+    fn new(fill: f64) -> Self {
+        Coor4D([fill; 4])
     }
-}
 
-impl IndexMut<usize> for Coor4D {
-    fn index_mut(&mut self, i: usize) -> &mut Self::Output {
-        &mut self.0[i]
+    fn dim(&self) -> usize {
+        4
     }
-}
 
-impl Add for Coor4D {
-    type Output = Self;
-    fn add(self, other: Self) -> Self {
-        Coor4D([
-            self.0[0] + other.0[0],
-            self.0[1] + other.0[1],
-            self.0[2] + other.0[2],
-            self.0[3] + other.0[3],
-        ])
+    fn nth_unchecked(&self, n: usize) -> f64 {
+        self.0[n]
     }
-}
 
-impl Add<&Coor4D> for Coor4D {
-    type Output = Self;
-    fn add(self, other: &Self) -> Self {
-        Coor4D([
-            self.0[0] + other.0[0],
-            self.0[1] + other.0[1],
-            self.0[2] + other.0[2],
-            self.0[3] + other.0[3],
-        ])
-    }
-}
-
-impl Sub for Coor4D {
-    type Output = Self;
-    fn sub(self, other: Self) -> Self {
-        Coor4D([
-            self.0[0] - other.0[0],
-            self.0[1] - other.0[1],
-            self.0[2] - other.0[2],
-            self.0[3] - other.0[3],
-        ])
-    }
-}
-
-impl Mul for Coor4D {
-    type Output = Self;
-    fn mul(self, other: Self) -> Self {
-        Coor4D([
-            self.0[0] * other.0[0],
-            self.0[1] * other.0[1],
-            self.0[2] * other.0[2],
-            self.0[3] * other.0[3],
-        ])
-    }
-}
-
-impl Div for Coor4D {
-    type Output = Self;
-    fn div(self, other: Self) -> Self {
-        Coor4D([
-            self.0[0] / other.0[0],
-            self.0[1] / other.0[1],
-            self.0[2] / other.0[2],
-            self.0[3] / other.0[3],
-        ])
+    fn set_nth_unchecked(&mut self, n: usize, value: f64) {
+        self.0[n] = value;
     }
 }
 
@@ -158,7 +98,7 @@ impl Coor4D {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::prelude::*;
+    use std::ops::{Add, Div, Mul};
 
     #[test]
     fn distances() {